Fix swapped test names
Unit test classes/files names for LimeEnumeratorRefsValidator and LimeFieldsValidator were swapped. Fixed the naming.
